some improvements
- converted original image to double such that rescaling for cell
search works
-  adaptions to active contour regarding the initialisation and update
of constants
- changed mask so that bright cells cannot influence the result (needs
to be further adapted for the edge images)
- small adjustment to table dimensions in measurement script

+++ input_output/Loaders/celltracks.m
@@ -127,7 +127,7 @@ classdef celltracks < loader
             rawImage = zeros(self.imageSize);
             for i=1:self.nrOfChannels;
                 try
-                    imagetemp = imread(self.imageFileNames{imageNr},i, 'info',self.tiffHeaders{imageNr});
+                    imagetemp = double(imread(self.imageFileNames{imageNr},i, 'info',self.tiffHeaders{imageNr}));
                 catch
                     notify(self,'logMessage',logmessage(2,['Tiff from channel ' num2str(ch) ' is not readable!'])) ;
                     return
